Break copyCartToOrder to it's own file
